add adjusted paths for duckdb comparison levels
