+++ splink/duckdb/duckdb_comparison_template_library.py
@@ -1,32 +1,6 @@
-from ..comparison_template_library import (
-    DateComparisonBase,
-    ForenameSurnameComparisonBase,
-    NameComparisonBase,
-    PostcodeComparisonBase,
+from .duckdb_helpers.duckdb_comparison_imports import (
+    date_comparison,
+    name_comparison,
+    forename_surname_comparison,
+    postcode_comparison,
 )
-from .duckdb_comparison_level_library import distance_function_level
-from .duckdb_comparison_library import DuckDBComparisonProperties
-
-
-class date_comparison(DuckDBComparisonProperties, DateComparisonBase):
-    @property
-    def _distance_level(self):
-        return distance_function_level
-
-
-class name_comparison(DuckDBComparisonProperties, NameComparisonBase):
-    @property
-    def _distance_level(self):
-        return distance_function_level
-
-
-class forename_surname_comparison(
-    DuckDBComparisonProperties, ForenameSurnameComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return distance_function_level
-
-
-class postcode_comparison(DuckDBComparisonProperties, PostcodeComparisonBase):
-    pass
